refactor enum choices
